requirements.txt
google.generativeai
streamlit
dotenv